Moved and renamed Visualizer.js
